feat(795): move widget to corresponding feature
