+++ src/features/student/ams/locales/en.json
@@ -67,3 +67,4 @@
     }
   }
 }
+
