+++ src/features/student/ams/locales/fr.json
@@ -67,3 +67,4 @@
     }
   }
 }
+
