+++ src/features/student/selfKnowledge/locales/en.json
@@ -113,3 +113,4 @@
     }
   }
 }
+
