+++ src/features/student/skills/components/cards/StudentSkillsWidget/StudentSkillsWidget.stub.ts
