+++ src/features/student/skills/components/cards/StudentSkillsWidget/StudentSkillsWidget.test.ts
@@ -4,7 +4,7 @@ import type { UseQueryDefinedReturnType } from '@tanstack/vue-query'
 import type { VueWrapper } from '@vue/test-utils'
 import type { Ref } from 'vue'
 import { mockedProgramsProgressOverview } from '@/__mocks__/fixtures/student'
-import StudentSkillsWidget from '@/features/student/skills/components/widgets/StudentSkillsWidget/StudentSkillsWidget.vue'
+import StudentSkillsWidget from '@/features/student/skills/components/cards/StudentSkillsWidget/StudentSkillsWidget.vue'
 import { useStudentCoursesSummaryQuery } from '@/features/student/skills/queries/use-skills-view.query/use-skills-view.query'
 import { BddTest } from '@avenirs-esr/avenirs-dsav/test-utils'
 import { mockAddErrorMessage } from 'tests/mocks'
