+++ src/features/student/traces/components/cards/StudentTracesWidget/StudentTracesWidget.stub.ts
