+++ src/features/student/traces/components/cards/StudentTracesWidget/StudentTracesWidget.test.ts
@@ -3,7 +3,7 @@ import { server } from '@/__mocks__/msw/server'
 import {
   StudentTraceCardStub
 } from '@/features/student/traces/components/cards/StudentTraceCard/StudentTraceCard.stub'
-import StudentTracesWidget from '@/features/student/traces/components/widgets/StudentTracesWidget/StudentTracesWidget.vue'
+import StudentTracesWidget from '@/features/student/traces/components/cards/StudentTracesWidget/StudentTracesWidget.vue'
 import { AvButtonStub, BddTest } from '@avenirs-esr/avenirs-dsav/test-utils'
 import { QueryClient, VueQueryPlugin } from '@tanstack/vue-query'
 import { flushPromises, type VueWrapper } from '@vue/test-utils'
