+++ src/features/student/traces/locales/fr.json
@@ -6,6 +6,23 @@
           "text": "0 trace | 1 trace | {count} traces"
         }
       },
+      "cards": {
+        "StudentTraceCard": {
+          "activityCount": "0 mise en situation | 1 mise en situation | {count} mises en situation",
+          "lifeProject": "Projet de vie",
+          "skillCount": "0 compétence | 1 compétence | {count} compétences",
+          "tagLabel": {
+            "group": "Groupe",
+            "solo": "Individuel"
+          }
+        },
+        "StudentTracesWidget": {
+          "buttons": {
+            "seeAll": "Voir toutes mes traces"
+          },
+          "title": "Mes dernières traces"
+        }
+      },
       "composites": {
         "TraceAssociations": {
           "caption": "Les associations de ma trace ({count})"
@@ -269,18 +286,8 @@
             }
           }
         }
-      },
-      "cards": {
-        "StudentTraceCard": {
-          "activityCount": "0 mise en situation | 1 mise en situation | {count} mises en situation",
-          "lifeProject": "Projet de vie",
-          "skillCount": "0 compétence | 1 compétence | {count} compétences",
-          "tagLabel": {
-            "group": "Groupe",
-            "solo": "Individuel"
-          }
-        }
       }
     }
   }
 }
+
