+++ src/features/student/user/locales/en.json
@@ -1,7 +1,88 @@
 {
   "student": {
     "user": {
-      
+      "cards": {
+        "StudentOverviewWidget": {
+          "bannerAlt": "Profile banner",
+          "buttons": {
+            "editProfile": "Edit my profile",
+            "establishments": "My establishment(s)",
+            "shareCofolio": "Share my Cofolio",
+            "shareResume": "Share my resume"
+          },
+          "pictureAlt": "Profile picture"
+        }
+      },
+      "overlays": {
+        "StudentMailboxPopover": {
+          "bodyNoNew": "You have no new messages to consult.",
+          "buttons": {
+            "close": "Close",
+            "navigate": "Go to my mailbox",
+            "seeAll": "See all"
+          },
+          "title": "No new messages | 1 unread message | {count} unread messages"
+        },
+        "StudentNotificationsPopover": {
+          "bodyNoNew": {
+            "assessedSkill": "A third party will have assessed you on a skill",
+            "comingUpEvent": "An event is coming up",
+            "header": "You will receive a notification in the following cases:",
+            "teacherMessage": "A teacher will send you a message",
+            "validatedTrace": "A trace has been validated"
+          },
+          "buttons": {
+            "close": "Close",
+            "seeAll": "See all"
+          },
+          "title": "No notification | 1 unread notification | {count} unread notifications"
+        },
+        "StudentProfilePopover": {
+          "buttons": {
+            "goToDw": "Go to my DW",
+            "goToSkills": "Go to the skills passport",
+            "logout": "Logout",
+            "manageProfile": "Manage my profile",
+            "seeCalendar": "See my calendar"
+          }
+        },
+        "UpdateProfileDrawer": {
+          "identity": {
+            "biography": {
+              "errors": {
+                "tooLong": "The biography must not exceed {maxLength} characters"
+              },
+              "label": "Biography"
+            },
+            "firstname": "First name",
+            "lastname": "Last name",
+            "mail": "Personal email address",
+            "title": "Identity"
+          },
+          "onDelete": {
+            "error": "An error occurred while deleting the photo.",
+            "success": "Your photo has been deleted."
+          },
+          "onUpdate": {
+            "error": "An error occurred while updating the profile.",
+            "success": "Your profile has been updated."
+          },
+          "pictures": {
+            "banner": "Banner",
+            "picture": "Profile picture"
+          },
+          "title": "Edit my profile"
+        }
+      },
+      "views": {
+        "StudentMailboxView": {
+          "title": "My mailbox"
+        },
+        "StudentNotificationsView": {
+          "title": "My notifications"
+        }
+      }
     }
   }
 }
+
