+++ src/features/student/user/locales/fr.json
@@ -1,7 +1,92 @@
 {
   "student": {
     "user": {
-      
+      "cards": {
+        "StudentOverviewWidget": {
+          "bannerAlt": "Bannière du profil",
+          "buttons": {
+            "editProfile": "Modifier mon profil",
+            "establishments": "Mon/Mes établissement(s)",
+            "shareCofolio": "Partager mon Cofolio",
+            "shareResume": "Partager mon CV"
+          },
+          "pictureAlt": "Photo de profil"
+        }
+      },
+      "overlays": {
+        "StudentMailboxPopover": {
+          "bodyNoNew": "Vous n'avez aucun nouveau message à consulter.",
+          "buttons": {
+            "close": "Fermer",
+            "navigate": "Aller à ma messagerie",
+            "seeAll": "Voir tout"
+          },
+          "title": "Aucun nouveau message | 1 message non lu | {count} messages non lus"
+        },
+        "StudentNotificationsPopover": {
+          "bodyNoNew": {
+            "assessedSkill": "Un tiers vous aura évalué sur une compétence",
+            "comingUpEvent": "Un événement a lieu prochainement",
+            "header": "Vous recevrez une notification dans les cas suivants :",
+            "teacherMessage": "Un enseignant vous enverra un message",
+            "validatedTrace": "Une trace a été validée"
+          },
+          "buttons": {
+            "close": "Fermer",
+            "seeAll": "Voir tout"
+          },
+          "title": "Aucune notification | 1 notification non lue | {count} notifications non lues"
+        },
+        "StudentProfilePopover": {
+          "buttons": {
+            "goToDw": "Aller sur mon ENT",
+            "goToSkills": "Aller sur le passeport de compétences",
+            "logout": "Me déconnecter",
+            "manageProfile": "Gérer mon profil",
+            "seeCalendar": "Voir mon agenda"
+          }
+        },
+        "UpdateProfileDrawer": {
+          "confirmDeletePhotoModal": {
+            "subtitle": "Cette action supprimera définitivement votre photo de profil. Confirmer la suppression ?",
+            "title": "Êtes-vous sûr de vouloir supprimer votre photo?"
+          },
+          "identity": {
+            "biography": {
+              "errors": {
+                "tooLong": "La biographie ne doit pas dépasser les {maxLength} charactères"
+              },
+              "label": "Biographie"
+            },
+            "firstname": "Prénom",
+            "lastname": "Nom",
+            "mail": "Adresse mail personnelle",
+            "title": "Identité"
+          },
+          "onDelete": {
+            "error": "Une erreur est survenue lors de la suppression de votre photo.",
+            "success": "Votre photo a été supprimée."
+          },
+          "onUpdate": {
+            "error": "Une erreur est survenue lors de la mise à jour du profil.",
+            "success": "Votre profil a été mis à jour."
+          },
+          "pictures": {
+            "banner": "Couverture",
+            "picture": "Photo de profil"
+          },
+          "title": "Modifier mon profil"
+        }
+      },
+      "views": {
+        "StudentMailboxView": {
+          "title": "Ma messagerie"
+        },
+        "StudentNotificationsView": {
+          "title": "Mes notifications"
+        }
+      }
     }
   }
 }
+
